Create shared calculated properties class
