+++ src/encoded/types/experiment.py
@@ -10,6 +10,10 @@ from .base import (
     paths_filtered_by_status,
 )
 from .dataset import Dataset
+from .shared_calculated_properties import (
+    CalculatedSlims,
+    CalculatedSynonyms
+)
 import datetime
 
 
@@ -20,7 +24,7 @@ import datetime
         'title': 'Experiments',
         'description': 'Listing of Experiments',
     })
-class Experiment(Dataset):
+class Experiment(Dataset, CalculatedSlims, CalculatedSynonyms):
     item_type = 'experiment'
     schema = load_schema('encoded:schemas/experiment.json')
     base_types = [Dataset.__name__] + Dataset.base_types
@@ -103,68 +107,6 @@ class Experiment(Dataset):
     rev.update({
         'replicates': ('Replicate', 'experiment')
     })
-
-    @calculated_property(condition='biosample_term_id', schema={
-        "title": "Organ slims",
-        "type": "array",
-        "items": {
-            "type": "string",
-        },
-    })
-    def organ_slims(self, registry, biosample_term_id):
-        if biosample_term_id in registry['ontology']:
-            return registry['ontology'][biosample_term_id]['organs']
-        return []
-
-    @calculated_property(condition='biosample_term_id', schema={
-        "title": "System slims",
-        "type": "array",
-        "items": {
-            "type": "string",
-        },
-    })
-    def system_slims(self, registry, biosample_term_id):
-        if biosample_term_id in registry['ontology']:
-            return registry['ontology'][biosample_term_id]['systems']
-        return []
-
-    @calculated_property(condition='biosample_term_id', schema={
-        "title": "Developmental slims",
-        "type": "array",
-        "items": {
-            "type": "string",
-        },
-    })
-    def developmental_slims(self, registry, biosample_term_id):
-        if biosample_term_id in registry['ontology']:
-            return registry['ontology'][biosample_term_id]['developmental']
-        return []
-
-    @calculated_property(condition='biosample_term_id', schema={
-        "title": "Biosample synonyms",
-        "type": "array",
-        "items": {
-            "type": "string",
-        },
-    })
-    def biosample_synonyms(self, registry, biosample_term_id):
-        if biosample_term_id in registry['ontology']:
-            return registry['ontology'][biosample_term_id]['synonyms']
-        return []
-
-    @calculated_property(condition='assay_term_id', schema={
-        "title": "Assay synonyms",
-        "type": "array",
-        "items": {
-            "type": "string",
-        },
-    })
-    def assay_synonyms(self, registry, assay_term_id):
-        if assay_term_id in registry['ontology']:
-            return registry['ontology'][assay_term_id]['synonyms'] + [
-                registry['ontology'][assay_term_id]['name'],
-            ]
-        return []
 
     @calculated_property(condition='date_released', schema={
         "title": "Month released",
