+++ src/encoded/types/shared_calculated_properties.py
@@ -1,0 +1,71 @@
+from contentbase import (
+    calculated_property
+)
+
+
+class CalculatedSlims:
+
+    @calculated_property(condition='biosample_term_id', schema={
+        "title": "Organ slims",
+        "type": "array",
+        "items": {
+            "type": "string",
+        },
+    })
+    def organ_slims(self, registry, biosample_term_id):
+        if biosample_term_id in registry['ontology']:
+            return registry['ontology'][biosample_term_id]['organs']
+        return []
+
+    @calculated_property(condition='biosample_term_id', schema={
+        "title": "System slims",
+        "type": "array",
+        "items": {
+            "type": "string",
+        },
+    })
+    def system_slims(self, registry, biosample_term_id):
+        if biosample_term_id in registry['ontology']:
+            return registry['ontology'][biosample_term_id]['systems']
+        return []
+
+    @calculated_property(condition='biosample_term_id', schema={
+        "title": "Developmental slims",
+        "type": "array",
+        "items": {
+            "type": "string",
+        },
+    })
+    def developmental_slims(self, registry, biosample_term_id):
+        if biosample_term_id in registry['ontology']:
+            return registry['ontology'][biosample_term_id]['developmental']
+        return []
+
+
+class CalculatedSynonyms:
+
+    @calculated_property(condition='biosample_term_id', schema={
+        "title": "Biosample synonyms",
+        "type": "array",
+        "items": {
+            "type": "string",
+        },
+    })
+    def biosample_synonyms(self, registry, biosample_term_id):
+        if biosample_term_id in registry['ontology']:
+            return registry['ontology'][biosample_term_id]['synonyms']
+        return []
+
+    @calculated_property(condition='assay_term_id', schema={
+        "title": "Assay synonyms",
+        "type": "array",
+        "items": {
+            "type": "string",
+        },
+    })
+    def assay_synonyms(self, registry, assay_term_id):
+        if assay_term_id in registry['ontology']:
+            return registry['ontology'][assay_term_id]['synonyms'] + [
+                registry['ontology'][assay_term_id]['name'],
+            ]
+        return []
